Create MarkovChain.py
This is a baseline strategy for predicting which machine the player will choose next. It is much less data-intensive than a neural network so will be easy to implement during this first week. ChatGPT wrote this code, but I went through all of it and made sure it makes sense.